pyproject.toml
[project]
name = "fastapi_items"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.24",
]

[tool.basedpyright]
typeCheckingMode = "basic"
reportExplicitAny = false
reportUnusedCallResult = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownArgumentType = false
reportUnknownLambdaType = false
reportUnknownParameterType = false
reportMissingTypeStubs = false
reportPrivateUsage = false
reportImportCycles = false
reportReturnType = false
